docs(zta/wc): move wc Go to examples dir
