+++ examples/workload-clusters/main.go
@@ -1,0 +1,316 @@
+package main
+
+import (
+	"bytes"
+	"context"
+	"encoding/json"
+	"fmt"
+	"log"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"time"
+
+	"github.com/gravitational/teleport/api/client"
+	"github.com/gravitational/teleport/api/client/proto"
+	headerv1 "github.com/gravitational/teleport/api/gen/proto/go/teleport/header/v1"
+	workloadcluster "github.com/gravitational/teleport/api/gen/proto/go/teleport/workloadcluster/v1"
+	"github.com/gravitational/teleport/api/types"
+)
+
+// TbotConfig defines a configuration for running tbot.
+type TbotConfig struct {
+	// Version is the configuration version.
+	Version string `json:"version"`
+	// Oneshot determines if tbot runs as a service.
+	Oneshot bool `json:"oneshot"`
+	// ProxyServer is the Teleport Proxy to run tbot against.
+	ProxyServer string `json:"proxy_server"`
+	// Onboarding defines how tbot should attempt to join the Teleport cluster.
+	Onboarding Onboarding `json:"onboarding"`
+	// Storage instructs tbot where to save its internal certificates.
+	Storage Storage `json:"storage"`
+	// Services defines which services for tbot to run.
+	Services []Service `json:"services"`
+}
+
+// Onboarding defines how tbot should attempt to join the Teleport cluster.
+type Onboarding struct {
+	// JoinMethod is how to join, such as iam.
+	JoinMethod string `json:"join_method"`
+	// Token is which token in the Teleport cluster to use.
+	Token string `json:"token"`
+}
+
+// Storage instructs tbot where to save its internal certificates.
+type Storage struct {
+	// Type is the storage type, such as "memory" for in-memory storage.
+	Type string `json:"type"`
+}
+
+// Service defines which services for tbot to run.
+type Service struct {
+	// Type is the service type, such as "identity".
+	Type string `json:"type"`
+	// Destination is used by the identity service to save retrieved identity files and certs.
+	Destination Destination `json:"destination"`
+}
+
+// Destination is used by the identity service to save retrieved identity files and certs.
+type Destination struct {
+	// Type is the type of storage, such as "path".
+	Type string `json:"type"`
+	// Path is the filepath to use.
+	Path string `json:"path"`
+}
+
+func main() {
+	ctx := context.Background()
+
+	/**********************************************
+	* Create a Teleport workload_cluster resource *
+	**********************************************/
+
+	// parentClient is a Teleport client connected to the parent cluster, which
+	// assumes the identity used has access for creating, reading, and deleting
+	// workload_cluster resources.
+	parentClient, err := client.New(ctx, client.Config{
+		Addrs: []string{
+			// This is the parent Teleport Cloud cluster's proxy address, such as:
+			"parent.teleport.sh",
+			// Note: port is optional.
+		},
+		Credentials: []client.Credentials{
+			// this loads the credential from tsh
+			client.LoadProfile("", ""),
+		},
+	})
+	if err != nil {
+		log.Panicf("failed to create client: %v", err)
+	}
+	defer parentClient.Close()
+
+	// wc defines a workload_cluster to create in us-west-2 with a bot named
+	// example-iam and configuration for a token to use IAM joining.
+	// The created child Teleport Cloud cluster will have a bot, role, and token
+	// each named example-iam. The bot will have access to create, read, and
+	// update users and roles.
+	wc := &workloadcluster.WorkloadCluster{
+		Kind:    types.KindWorkloadCluster,
+		Version: "v1",
+		Metadata: &headerv1.Metadata{
+			Name: "example",
+		},
+		Spec: &workloadcluster.WorkloadClusterSpec{
+			Regions: []*workloadcluster.Region{
+				{
+					Name: "us-west-2",
+				},
+			},
+			Bot: &workloadcluster.Bot{
+				Name: "example-iam",
+			},
+			Token: &workloadcluster.Token{
+				JoinMethod: "iam",
+				Allow: []*workloadcluster.Allow{
+					{
+						AwsAccount: "account",
+						AwsArn:     "arn",
+					},
+				},
+			},
+		},
+	}
+
+	// Create a workload_cluster resource within the parent Teleport Cloud cluster.
+	if _, err := parentClient.CreateWorkloadCluster(ctx, wc); err != nil {
+		log.Panicf("failed to create workload cluster: %v", err)
+	}
+
+	// Wait for the created workload cluster to reach an active state.
+	timeoutctx, cancel := context.WithTimeout(ctx, 10*time.Minute)
+	defer cancel()
+	wc, err = waitForActiveWorkloadCluster(timeoutctx, parentClient, wc.Metadata.Name, 30*time.Second)
+	if err != nil {
+		log.Panicf("failed waiting for workload cluster to be active: %v", err)
+	}
+
+	/************************************************
+	* Run tbot against child Teleport Cloud cluster *
+	************************************************/
+
+	// Create a directory that will be used for tbot's configuration and saving a
+	// retrieved identity file for interacting with the child Teleport Cloud cluster.
+	tbotDir, err := os.MkdirTemp("", "")
+	if err != nil {
+		log.Panicf("error creating directory for tbot: %v", err)
+	}
+	defer os.RemoveAll(tbotDir)
+
+	// Create the tbot configuration.
+	tbotConfig := TbotConfig{
+		// Version must be v2.
+		Version: "v2",
+		// Oneshot should be true to avoid running tbot as a daemon.
+		Oneshot: true,
+		// ProxyServer should be the Proxy Server including the port 443 for the
+		// new child Teleport Cloud cluster.
+		ProxyServer: fmt.Sprintf("%s:443", wc.Status.Domain),
+		Onboarding: Onboarding{
+			// Only iam join method will be supported in the short term for workload_clusters.
+			JoinMethod: "iam",
+			// Token must match the same name provided in the workload_cluster's
+			// Spec.Bot.Name.
+			Token: "example-iam",
+		},
+		Storage: Storage{
+			// Configure tbot to use in-memory storage.
+			Type: "memory",
+		},
+		Services: []Service{
+			{
+				Type: "identity",
+				Destination: Destination{
+					Type: "directory",
+					// A file named identity will be created in the provided path.
+					// This identity file may be provided to tctl or Teleport clients
+					// for interacting with a Teleport cluster.
+					Path: tbotDir,
+				},
+			},
+		},
+	}
+
+	// Write the tbot configuration to a `tbot.json` file.
+	tbotConfigContent, err := json.Marshal(tbotConfig)
+	if err != nil {
+		log.Panicf("error marshalling tbot configuration: %v", err)
+	}
+	tbotConfigPath := filepath.Join(tbotDir, "tbot.json")
+	if err := os.WriteFile(tbotConfigPath, tbotConfigContent, 0600); err != nil {
+		log.Panicf("error writing tbot configuration: %v", err)
+	}
+
+	// Run the tbot binary. Teleport does not expose programmatic access to
+	// tbot, so the binary must be used.
+	// Once tbot start has successfully completed then an identity file
+	// will be populated at the provided path in the tbot configuration.
+	var bufErr bytes.Buffer
+	tbotCmd := exec.Command("tbot", "start", "-c", tbotConfigPath)
+	tbotCmd.Stderr = &bufErr
+	if err := tbotCmd.Run(); err != nil {
+		log.Panicf("error running tbot: %v\n\n%s", err, bufErr.String())
+	}
+
+	/*************************************************************
+	* Manage roles and users in the child Teleport Cloud cluster *
+	*************************************************************/
+
+	// Create a new Teleport client to interact with the child Teleport Cloud cluster.
+	// This client will use the identify file retrieved by tbot.
+	childClient, err := client.New(ctx, client.Config{
+		Addrs: []string{
+			// This is the child Teleport Cloud cluster's proxy address:
+			wc.Status.Domain,
+			// Note: port is optional.
+		},
+		Credentials: []client.Credentials{
+			// This uses an identity file instead of using a credential from tsh.
+			client.LoadIdentityFile(filepath.Join(tbotDir, "identity")),
+		},
+	})
+	if err != nil {
+		log.Panicf("failed to create client: %v", err)
+	}
+	defer childClient.Close()
+
+	// The following section includes examples of creating and deleting
+	// a role and a user.
+
+	// Create a new role named example in the child Teleport Cloud cluster.
+	newRole := types.RoleV6{
+		Metadata: types.Metadata{
+			Name: "example",
+		},
+		Spec: types.RoleSpecV6{
+			Allow: types.RoleConditions{
+				Rules: []types.Rule{
+					{
+						Resources: []string{
+							"workload_cluster",
+						},
+						Verbs: []string{
+							"read",
+						},
+					},
+				},
+			},
+		},
+	}
+	if _, err := childClient.CreateRole(ctx, &newRole); err != nil {
+		log.Panicf("error creating role: %v", err)
+	}
+
+	// Create user named "example" that has the new "example" role assigned.
+	newUser := types.UserV2{
+		Metadata: types.Metadata{
+			Name: "example",
+		},
+		Spec: types.UserSpecV2{
+			Roles: []string{
+				"example",
+			},
+		},
+	}
+	if _, err := childClient.CreateUser(ctx, &newUser); err != nil {
+		log.Panicf("error creating user: %v", err)
+	}
+
+	// create an invite URL for user to activate account and setup MFA
+	resetPasswordToken := proto.CreateResetPasswordTokenRequest{
+		Name: newUser.Metadata.Name,
+		TTL:  proto.Duration(2 * time.Hour),
+		Type: "invite",
+	}
+	resetToken, err := childClient.CreateResetPasswordToken(ctx, &resetPasswordToken)
+	if err != nil {
+		log.Panicf("error creating reset token: %v", err)
+	}
+
+	ttl := resetToken.Expiry().Sub(time.Now().UTC())
+	log.Printf("User %q has been created but requires a password. Share this URL with the user to complete user setup, link is valid for %v:\n%v\n\n", newUser.Metadata.Name, ttl, resetToken.GetURL())
+
+	/****************************
+	* Delete a workload cluster *
+	****************************/
+
+	// Clean up the previously created workload_cluster resource in the parent
+	// Teleport Cloud cluster.
+	if err := parentClient.DeleteWorkloadCluster(ctx, wc.Metadata.Name); err != nil {
+		log.Panicf("error deleting cloud cluster: %v", err)
+	}
+}
+
+func waitForActiveWorkloadCluster(ctx context.Context, client *client.Client, workloadClusterName string, pollingInterval time.Duration) (*workloadcluster.WorkloadCluster, error) {
+	for {
+		timer := time.NewTimer(pollingInterval)
+
+		select {
+		case <-timer.C:
+			wc, err := client.GetWorkloadCluster(ctx, workloadClusterName)
+			if err != nil {
+				return nil, fmt.Errorf("error getting workload cluster: %w", err)
+			}
+
+			if wc.Status == nil {
+				continue
+			}
+
+			if wc.Status.State == "active" {
+				return wc, nil
+			}
+		case <-ctx.Done():
+			return nil, ctx.Err()
+		}
+	}
+}
